refactor(e2e): extrai utilitários Keycloak para lib shared-e2e e corrige #TODO
[S1] Move keycloak-login.ts (resetPasswords, keycloakLogin, keycloakLogout,
expectUserInHeader) de selecao-e2e e portal-e2e para libs/shared-e2e —
lib Nx com tag scope:shared, type:e2e. Ambos os apps passam a re-exportar
de @uniplus/shared-e2e, eliminando 131 linhas duplicadas. Quando
ingresso-e2e precisar de auth E2E, importa do mesmo ponto.

[S2] Substitui 'issue #TODO' por 'issue uniplus-web#84' nos três
environment.prod.ts (selecao, ingresso, portal).
