+++ libs/shared-e2e/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "outDir": "../../dist/out-tsc",
+    "strict": true
+  },
+  "include": ["src/**/*.ts"]
+}
